LayoutTests/media/video-src-set.html

<video controls></video>
<div>
Test that setting src attribute triggers load
</div>
<script src=video-test.js></script>
<script>
video.setAttribute('src', 'content/test.mp4')
waitForEventTestAndEnd('begin', "relativeURL(video.currentSrc) == 'content/test.mp4'");
</script>
